remove templating, move stuff around
