+++ app/Onboard/index.dev.html
@@ -1,0 +1,25 @@
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      http-equiv="Content-Security-Policy"
+      content="
+        script-src 'strict-dynamic' 'nonce-2f0d956b-0d9c-4f1e-874a-57a1ec828872' http://localhost:1234 'unsafe-eval';
+        base-uri 'self';
+        connect-src *;
+        img-src *;
+        frame-src 'none';
+        style-src 'self' http://localhost:1234 'unsafe-inline';
+        object-src 'none';
+      "
+    />
+    <title>Onboard</title>
+    <link rel="stylesheet" href="../../resources/base.styl" type="text/css" />
+    <link rel="stylesheet" href="../../app/Onboard/index.styl" type="text/css" />
+  </head>
+  <body>
+    <div class="frameOverlay"></div>
+    <div id="onboard"></div>
+    <script type="module" src="./index.js" nonce="2f0d956b-0d9c-4f1e-874a-57a1ec828872"></script>
+  </body>
+</html>
