--- renderer/index.ejs
@@ -1,26 +1,0 @@
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="
-        script-src 'strict-dynamic' 'nonce-2f0d956b-0d9c-4f1e-874a-57a1ec828872'<% if (this.isDev) { %> http://localhost:1234 'unsafe-eval'<% } %>;
-        base-uri 'self';
-        connect-src *;
-        img-src <% if (this.window === 'dapp') { %>blob: http://localhost:8421<% } else { %>*<% } %>;
-        frame-src 'none';
-        style-src 'self'<% if (this.isDev) { %> http://localhost:1234 'unsafe-inline'<% } %>;
-        object-src 'none';<% if (!this.isDev) { %>
-        require-trusted-types-for 'script';<% } %>
-      "
-    />
-    <title><%= this.windowTitle %></title>
-    <link rel="stylesheet" href="../../resources/base.styl" type="text/css" />
-    <link rel="stylesheet" href="../../app/<%= this.windowTitle %>/index.styl" type="text/css" />
-  </head>
-  <body>
-    <div class="frameOverlay"></div>
-    <div id="<%= this.window %>"></div>
-    <script type="module" src="./index.js" nonce="2f0d956b-0d9c-4f1e-874a-57a1ec828872"></script>
-  </body>
-</html>
